:recycle: refactor: 각종 쿼리 서비스 메소드 분리
해당 서비스에는 projectIsFull, getProject 메서드만

Related to: #23
